Modify KURL::SetPort to return if it successfully parses String
This changes KURL::SetPort from having no return value to return a bool
that is true when it successfully parses the String port and false if
it cannot parse the String port.

This is a precursor to addressing interop issues with URLPattern. See
bug https://crbug.com/409350827 for more information.

This also updates the existing SetPort unit test to validate the return
value.

Bug: 409350827
Change-Id: I91eecd57b375048a08529e3e6a8880462b609400
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6814731
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Shunya Shishido <sisidovski@chromium.org>
Commit-Queue: Dave Risney <david.risney@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1496684}