Simplify the logic for adding the module maps to the command line.
This ensures that the compiler flags for modules are provided iff a
target depends on "all_modules".

This allows us to safely remove several hacks that were in place to fix
some compile errors in the build. It also will improve on autogenerated
modules in the future, as mac can get quite horrifying with hundreds of
modules.

Bug: 425527533

Change-Id: I6a6a696449a0b1daf72a5f4fdf791b23807ee78e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6821092
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Commit-Queue: Matt Stark <msta@google.com>
Auto-Submit: Matt Stark <msta@google.com>
Cr-Commit-Position: refs/heads/main@{#1496743}