spanification: automatically spanify various pointers
This is the result of running the automatic spanification on linux and
updating code to use and pass spans where size is known.
This CL has no behavior changes.

All changes should be correct and no behaviour changes in patchset 1, however the plan is to follow up and remove the redundant size parameters. To this effect in patchset 2
we've added CHECKS and renames to note the fact the size is believed to be redundant. This was done through AI and should be carefully checked, but the CHECKS are not fatal and thus if unsure should be fine to land and remove after reports of the crashes.

The original patch was fully automated using script:
//tools/clang/spanify/rewrite-multiple-platforms.sh -platforms=linux

and then split with
`git cl split --target-range 8 32 --max-depth=2 --no-reviewers -d ~/description.txt`

This covers directories:
gpu

This is part of auto spanification see internal doc about it:
https://docs.google.com/document/u/1/d/1poEhNARfFas2osqPoXGzVZtBgnOQGQtQNUFTA-MlADs/edit?tab=t.0

This CL was uploaded by an experimental version of git cl split
(https://crbug.com/389069356).

Bug: 431824299
Change-Id: Ic88e2a6048438645e0876c68be590ea1bb124117
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6731992
Commit-Queue: Stephen Nusko <nuskos@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Auto-Submit: Stephen Nusko <nuskos@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1496769}