[FPF-CI] Refactor tracking of triggering canvas operations
Instead of capturing all (possibly triggering) canvas operations in a
context and CanvasPath, this CL aims to capture the triggering canvas
operations that were made in an accelerated context. This makes keeping
taint tracking a lot easier, and mitigates a number of issues (e.g.
drawImage of an accelerated canvas into an unaccelerated one should
still trigger an intervention). Furthermore, this removes the dependency
on the RenderingContext, which might not always be available (e.g. when
transferring to an OfflineContext).

Bug: 377325952
Change-Id: Ic6b74a4e5a2011229ad60e613b0c6d36a928bb7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6765542
Reviewed-by: Jean-Philippe Gravel <jpgravel@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Tom Van Goethem <tov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1496860}