[iOS] Fix setup-gn.py to fix target source map without using lldbinit.py
This CL aims to make it easier to work with several Chromium checkouts
on iOS, with one Xcode window opened for each checkout.

Problem: when an Xcode window is opened in checkout A and the debugger
is started, .lldbinit is loaded and the module defined by lldbinit.py is
imported, which sets the appropriate target.source-map for checkout A.
But then, when opening an Xcode window in checkout B and starting the
debugger in checkout B, although .lldbinit from checkout B is loaded, it
appears that running `import lldbinit` does not reload the module (even
though checkout B has its own lldbinit.py) so the value of the
target.source-map setting is not changed accordingly.

Fix: .lldbinit now no longer depends on lldbinit.py and directly sets
the `target.source-map` setting to the desired value. The environment
variable `CHROMIUM_LLDBINIT_SOURCED` is also still set so the behaviour
does not change in that regard.

Change-Id: If7c8d8745bdc5756799dcb249fea30eaf9d82a79
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6819057
Auto-Submit: Quentin Pubert <qpubert@google.com>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1496861}