[Autofill] Represent field types as list in BuildFieldDataLogs()
Since recently, AutofillField::Type() may be a set of FieldTypes.

Before this CL, "fieldType" in the Autofill Feedback JSON is a string,
which may be, for example, "ADDRESS_HOME_ZIP, PASSPORT_NUMBER".

After this CL, "fieldType" is not set anymore and instead "fieldTypes"
is a list of strings such as ["ADDRESS_HOME_ZIP", "PASSPORT_NUMBER"].

Bug: 40253237, 432645177
Change-Id: I41a476747ad25afcfa7f3d4dbac07db7f4365b02
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6819024
Commit-Queue: Christoph Schwering <schwering@google.com>
Reviewed-by: Norge Vizcay <vizcay@google.com>
Cr-Commit-Position: refs/heads/main@{#1496919}