Switch variations protos to use edition syntax.
This brings them closer to the syntax used in the server-side,
google3 versions, although other differences remain. This was done
by copying over the internal versions and then manually inspecting
the diff to remove many changes unrelated to the syntax conversion,
such as internal fields and comments. Other comments that seemed
to have benign differences are simply updated.

I only covered the protos used as part of the variations seed
delivery, omitting changes to client_variations.proto and
stored_seed_info.proto. These can be handled separately.

However client_variations.proto is currently compiled into a
JS target and the JS proto compiler doesn't support edition
syntax. However, the plan is to remove that dependency -
see crbug.com/40253708. This CL simply moves it to its own
target.

One benefit of the edition syntax is to be able to more
finely control the generated code output at a file and field
level. For example, in this change, various semantics of
proto2 are preserved, such as:
  1. features.enum_type = CLOSED, specifying that unknown
     enum values result in the default value for that enum
     being set.
  2. features.utf8_validation = NONE, specifying that no
     validation is done on utf8.
  3. features.field_presence = LEGACY_REQUIRED on some
     fields, that specifies they use the old "required"
     logic for those fields
  4. features.repeated_field_encoding = EXPANDED on repeated
     fields, specifying they shouldn't be packed.

Note: Although consolidating the syntaxes is a benefit in
itself, my motivation for this change is actually to be able
to customize 1. above to use OPEN enum types for better control
during validation of studies, which I plan to revamp as part of
adding the sticky study activation type that I'm working on.

Bug: 435630455
Change-Id: If578af49f2781830ef591f103f5d4e114827e9b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6819102
Commit-Queue: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Ramon Cano Aparicio <rcanoaparicio@google.com>
Cr-Commit-Position: refs/heads/main@{#1496920}