[Autofill] Return FieldType representation in AutofillType::ToString()
Before this CL, for an HtmlFieldType t, AutofillType(t).ToString()
returned the string representation of the HtmlFieldType.

After this CL, it returns the string representation of
HtmlFieldTypeToBestCorrespondingFieldType(t).

AutofillType::ToString() is only used for logging:
- autocomplete-information attribute
- chrome://autofill-internals
Both also contain the serialization of the HtmlFieldType (if the field
has one).

This change is necessary to eliminate HtmlFieldType from AutofillType.

Bug: 436013479, 432645177
Change-Id: Ib4a3edf17c02adad271ab1f0ee59b62638ba9533
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6818288
Reviewed-by: Jan Keitel <jkeitel@google.com>
Commit-Queue: Christoph Schwering <schwering@google.com>
Cr-Commit-Position: refs/heads/main@{#1496950}