Fix checking for dev-mode status
Enable Puppeteer tests to use CDP methods on
ChromeOS if dev-mode is enabled.
is_official_build isn't the way to check if the
device is in dev-mode, since it's a runtime thing.
Added a command-line switch instead of feature flag, so that it cannot be set unless the device is in dev-mode as well.

TEST=tested locally on a ChromeOS VM with "autoninja -C out_betty/Release/ chrome && ./third_party/chromite/bin/deploy_chrome --build-dir=out_betty/Release --device=localhost:9222 --compress=always"

Bug: b:436210591
Change-Id: I233845b228605f0ec832e4d53c05fc1e0c81dd0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6810909
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Mohamed Omar <mohamedaomar@google.com>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1496953}