Change latest_serial_number_ type to optional.
In order to get the seed serial number we need to read the seed
contents, which is an expensive operation. To avoid this, the
serial number is read the first time and then stored into memory.

If no seed is fetched, latest_serial_number_ will be empty and it
will parse the seed each time.

Change the variable type to optional<string> to differentiate
between absence of value and empty string.

Change-Id: Iae32401a706ff698d6c60f920da59c20bf15540d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6811047
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Commit-Queue: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1496958}