[hub] Enable the search loupe on tab groups pane for hub search
Enable the search loupe of the tab groups pane for hub search, through
decoupling the menu button from the search loupe. The visibility used
to be controlled via the container for the 3-dot menu button but will
now be controlled via the button wrapper.

Bug: 436529097
Change-Id: Ia4022c1d60e0db4e99765bba44880f623a9a0882
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6819264
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Commit-Queue: Brandon Fong <bjfong@google.com>
Cr-Commit-Position: refs/heads/main@{#1496961}