[Autofill] Remove HtmlFieldType from AutofillType
This CL replaces AutofillType::html_type() with a Boolean
AutofillType::is_country_code(). The semantics of this Boolean is to
indicate whether ADDRESS_HOME_COUNTRY is a country code.

It replaces code like
  AutofillType(HtmlFieldType::kCountryCode)
with
  AutofillType(ADDRESS_HOME_COUNTRY, /*is_country_code=*/true)

The purpose of this change is to prevent new dependencies on
AutofillType::html_type().

In particular, it allows us to remove the assumption in
AutofillField::MakeAutofillType() that the references to
AutofillType::html_type() only care about HtmlFieldType::kCountryCode.

Bug: 436013479, 432645177
Change-Id: I9267bc4d2f1bf5ab15b56636e627e57b1e6f06df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6818462
Commit-Queue: Christoph Schwering <schwering@google.com>
Reviewed-by: Jan Keitel <jkeitel@google.com>
Cr-Commit-Position: refs/heads/main@{#1496967}