RenderDocument: Add about:blank deferred test.
The existing ReuseSameRFHNotDeferred test is checking that an RFH will
be reused when navigating with the same RFH from about:blank to a URL.
Since the whole point of RenderDocument is to not reuse RFHs, this test
will no longer be useful when RenderDocument is enabled.

To still have coverage for the about:blank case, this CL adds a new test
for the RenderDocument enabled case that checks that a speculative RFH
is created after the navigation request is sent.

Bug: 40615943
Change-Id: Ibb03290a714e26f2d05d174c22b514ff154a03d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6813820
Reviewed-by: Jiacheng Guo <gjc@google.com>
Commit-Queue: Liam Brady <lbrady@google.com>
Reviewed-by: Charlie Reis <creis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1496969}