[PWA/AppUpdating]: Fix icon writing behavior for empty app bitmaps
Previously, writing the pending update bitmaps to the disk would cause
the app's icons to get removed. This is because writing the app's icon
bitmaps would happen regardless of whether they were populated, and if
the passed icons were empty, the app icons would be removed based on an
implementation detail of AtomicallyWriteIcons().

This CL fixes that behavior by ensuring that app icons are written ONLY
if the passed bitmaps are populated and valid, which matches the
behavior for all other types of icon writes here. Also adds tests to
verify that this works well with icons that are not the primary product
icons, like shortcut menu icons.

Bug: 428973284
Change-Id: If5dbf43699eb886a94e3f966c680cb19327abb24
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6806377
Reviewed-by: Dibyajyoti Pal <dibyapal@chromium.org>
Commit-Queue: May Siem <msiem@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1496990}