Add a README to components/history/core
There are three goals for this CL: (1) Add more context to the code
contained in the core/ directory. (2) Inform users of the layering
constraints and cross-platform nature of the History code. (3) Serve as
a landing point for developers who are coming across unknown
ios-simulator trybot failures due to layering issues.

Change-Id: I7faf5b65f775c8006428fbe117f214aab8bcfec9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6818037
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Commit-Queue: Kyra Seevers <kyraseevers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1497015}