Migrate EndpointFetcher usage to preferred RequestParams constructor
The number of parameters being passed to the EndpointFetcher is
becoming difficult to manage. As such, the EndpointFetcher has
been modified so Request Parameters are passed in their own struct
which uses the builder pattern. This CL migrates one of the
EndpointFetcher usages in components/mantra to use the new constructor.

Bug: 435743536
Change-Id: Idabff338989b90189651e493ea8f485d0604312f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6812942
Reviewed-by: Michael Crouse <mcrouse@chromium.org>
Commit-Queue: David Maunder <davidjm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1497058}