[TabStripService api] Add C++ typemap for ImageSkia
WebUI uses a data uri to process images like the fav icon. In C++ -
specifically for Views, there is no reason to transform an image to a
uri and then back to an image.

This CL typemaps the image (specifically an ImageSkia) to avoid the
conversation step.

Bug: 435491236
Change-Id: I85bcf46c0b42b39062cb8d39f223621d4ca41520
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6808516
Commit-Queue: David Yeung <dayeung@chromium.org>
Reviewed-by: Fred Shih <ffred@chromium.org>
Reviewed-by: Tom Lukaszewicz <tluk@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1497101}