crash: remove temporary command output from minidumps
This is a partial revert of f6ee10c0369d
(https://chromium-review.googlesource.com/c/6788541). The base/process
portions of that change were intended to be permanent and are not
reverted.

f6ee10c0369d was very successful in further diagnosing
https://crbug.com/40064248, with complete notes on that bug at comments
117 and 120, and conclusions at comments 118, 119, 121, and 123–125. In
particular, comment 119 documents a reliable reproduction.

The new information is sufficient to file a new Feedback report with
Apple, now on file as FB19384824.

The new debugging information is no longer necessary. It’s served its
purpose well and now it’s time to remove it.

Bug: 40064248
Change-Id: Ie80e6a0844deb5aed24e5d97a8f23ae9bd2fa72c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6818049
Commit-Queue: Mark Mentovai <mark@chromium.org>
Reviewed-by: David Benjamin <davidben@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1497137}