[PinnedTabs] Strip pinned status of tab while adding to a new group
A tab should not be simultaneously pinned and in a tab group. This change ensures that whenever a tab is added to a group, its pinned status is automatically removed.

This is achieved by calling a new `unpinTabIfNeeded` helper method from all the code paths that handle adding tabs to groups.

Bug: 427929717
Change-Id: I2ce1e0a8de8f5d2b969d13046a2d1c3b7a8e8877
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6820914
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Commit-Queue: Madhav Pruthi <madhavpruthi@google.com>
Cr-Commit-Position: refs/heads/main@{#1497164}