[Compose] Toggle disclaimer strings based on enterprise policy
New strings are based on the mocks provided in the bug, which are confirmed to be the source of truth. An additional piece of this change is updating the support link for enterprise users from the Compose learn more page to the enterprise support page for "Chrome—Generative AI features and policies".

Bug: b:374323410
Change-Id: Ic49cad4b6138bfd2e18f14a72758145456527ffc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6700275
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Reviewed-by: Trevor Perrier <perrier@chromium.org>
Commit-Queue: Anthony Cui <cuianthony@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1497231}