NTP Promos: Make "has promos" check consistent
The original check considered only promo elgibility when determining if
there are promos to show. The actual generation of promo lists can
suppress promos for several reasons, so rather than constructing a web
component that might not have anything to show, just align the two
methods.

Efficiency is not a concern here with a small number of promos, but
conditionally skip ordering the list (the heaviest part) for peace
of mind.

Bug: 433606749
Change-Id: Ie2852dcddb1eb2b2f09cc5e8f54bc9d7356a8acd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6806480
Auto-Submit: Christopher Grant <cjgrant@chromium.org>
Reviewed-by: David Pennington <dpenning@chromium.org>
Commit-Queue: Christopher Grant <cjgrant@chromium.org>
Reviewed-by: Kaan Alsan <alsan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1497258}