[CLAMS] Propagate GenerateOptions to Android Session
There isn't a generic mojo utility for converting c++ mojo struct to
java[1], so disassemble the struct in c++ and reassemble it in java.
Only passing the parameters that are supported on Android (max_output_tokens). max_output_tokens is equivalent to targetReplyLengthMax in AiCore.

[1]
https://chromium.googlesource.com/chromium/src/mojo/+/HEAD/public/java/bindings/README.md#missing-type-conversions-structtraits

Bug: 425408635
Change-Id: I413ebbda254bb1d9a719652b57968631e65cdff6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6819592
Commit-Queue: Xinghui Lu <xinghuilu@chromium.org>
Reviewed-by: Clark DuVall <cduvall@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1497268}