Add a default case for all switch statements in ts
Starting with ts, because the change is already quite substantial.

The styleguide requires that all switch statements *must* have a default case.
We would like to turn on the linter for this styleguide, but need to fix all
the missing cases first. Styleguide:

https://google.github.io/styleguide/tsguide.html#switch-statements

For cases where it was just falling through, we used:
  default:
    break;

For cases where it was returning something (that cannot be undefined),
we used:
  default:
    assertNotReached();

This change should be a noop change.

NO_IFTTT=styleguide change, no modification to underlying enum

This CL was uploaded by git cl split.

R=wesokuhara@google.com

Bug: 436274263
Change-Id: Id8ccf19da2333e6cf98a1eee3302e44b33896e38
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6822450
Commit-Queue: Wes Okuhara <wesokuhara@google.com>
Auto-Submit: Fred Shih <ffred@chromium.org>
Reviewed-by: Wes Okuhara <wesokuhara@google.com>
Cr-Commit-Position: refs/heads/main@{#1497306}