Add a default case for all switch statements in ts
Starting with ts, because the change is already quite substantial.

The styleguide requires that all switch statements *must* have a default
case. We would like to turn on the linter for this styleguide, but need
to fix all the missing cases first. Styleguide:

https://google.github.io/styleguide/tsguide.html#switch-statements

For cases where it was just falling through, we used:
  default:
    break;

For cases where it was returning something (that cannot be undefined),
we used:
  default:
    assertNotReached();

This change should be a noop change.

NO_IFTTT=styleguide change, no modification to underlying enum

This CL was uploaded by git cl split.

R=thestig@chromium.org

Bug: 436274263
Change-Id: I8ceb0a4af0e07f2e8d3cb39881b06a9c7715868d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6822550
Reviewed-by: Lei Zhang <thestig@chromium.org>
Auto-Submit: Fred Shih <ffred@chromium.org>
Commit-Queue: Fred Shih <ffred@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1497319}