Add a default case for all switch statements in ts
Starting with ts, because the change is already quite substantial.

The styleguide requires that all switch statements *must* have a default case.
We would like to turn on the linter for this styleguide, but need to fix all
the missing cases first. Styleguide:

https://google.github.io/styleguide/tsguide.html#switch-statements

For cases where it was just falling through, we used:
  default:
    break;

For cases where it was returning something (that cannot be undefined),
we used:
  default:
    assertNotReached();

This change should be a noop change.

NO_IFTTT=styleguide change, no modification to underlying enum

This CL was uploaded by git cl split.

R=hujasonx@google.com

Bug: 436274263
Change-Id: Ie6209ad31d541ca8c6ded1af803c258823715b73
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6822756
Reviewed-by: Jason Hu <hujasonx@google.com>
Commit-Queue: Jason Hu <hujasonx@google.com>
Auto-Submit: Fred Shih <ffred@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1497321}