Fix segfault in crashpad handler on ARM64 Android
On some ARM64 Android devices, the crashpad handler would segfault on
startup. This manifested as a missing crash report on chrome://crashes
after triggering a crash via chrome://crash. This was caused by a symbol
conflict between the version of protobuf used by Chromium and the
version of protobuf used by the system.

The crashpad handler has a transitive dependency on the systems
libandroid.so, which in turn has a transitive dependency on the systems
libprotobuf-cpp-lite.so. This caused both versions of protobuf to be
loaded into the same process, which resulted in a crash during static
initialization.

This change fixes the crash by statically linking the Chromium version
of protobuf into the crashpad handler. This resolves the symbol
conflict. This change has been verified to fix the issue.

This CL makes protobuf_lite to always use static_library on android.

The size of ChromePublic.apk on component builds increases from 737M to
742M.

Bug: 403492235
Change-Id: I2d3a1f2a5d154db8601505f62f3ce96a0c184a50
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6804724
Commit-Queue: Keigo Oka <oka@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Auto-Submit: Keigo Oka <oka@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1497324}